+++ src/Podenv/Image.hs
@@ -8,12 +8,10 @@ module Podenv.Image (
     mkImageName,
     nixCommandProfile,
     nixCommandPath,
-    nixFlags,
     nixArgs,
 ) where
 
 import Data.Digest.Pure.SHA qualified as SHA
-import Data.Text qualified as Text
 import Podenv.Dhall
 import Podenv.Prelude
 
@@ -38,38 +36,8 @@ nixCommandPath = "/nix/" <> nixCommandProfile <> "/bin/nix"
 nixFlags :: [Text]
 nixFlags = ["--extra-experimental-features", "nix-command flakes"]
 
-defaultNixCache, defaultNixCachePublicKey :: Text
-defaultNixCache = "https://cache.nixos.org"
-defaultNixCachePublicKey = "cache.nixos.org-1:6NCHdD59X431o0gWypbMrAURkbJ16ZPMQFGspcDShjY="
-
-{- | Get the cache url from its public key
-
->>> cacheUrlFromKey "podenv.cachix.org-1:FA80Dv5XSHxzMYOg1wEANhDN7h43nN8twBJiYNzcDGY="
-"https://podenv.cachix.org"
--}
-cacheUrlFromKey :: Text -> Text
-cacheUrlFromKey cacheKey = case rest of
-    "" -> error $ ".cachix.org-1 not found in flake cache: '" <> cacheKey <> "'"
-    _ -> "https://" <> cacheUrlPrefix <> ".cachix.org"
+nixArgs :: Text -> [Text]
+nixArgs installable = impureArgs <> nixFlags <> [installable]
   where
-    (cacheUrlPrefix, rest) = Text.breakOn ".cachix.org-1:" cacheKey
-
-nixArgs :: Flakes -> [Text]
-nixArgs flakes = impureArgs <> nixExtraArgs <> nixCacheArgs <> installables flakes
-  where
-    impureArgs
-        | any (Text.isInfixOf "nix-community/nixGL") (installables flakes) = ["--impure"]
-        | otherwise = []
-    nixCacheArgs = case cache flakes of
-        Just cacheKey ->
-            [ "--option"
-            , "binary-caches"
-            , Text.unwords [defaultNixCache, cacheUrlFromKey cacheKey]
-            , "--option"
-            , "trusted-public-keys"
-            , Text.unwords [defaultNixCachePublicKey, cacheKey]
-            ]
-        Nothing -> []
-    nixExtraArgs = case nixpkgs flakes of
-        Just pin | not (all (Text.isPrefixOf pin) (installables flakes)) -> ["--override-input", "nixpkgs", pin]
-        _ -> []
+    -- impure is necessary for nixGL to discover the nvidia driver
+    impureArgs = ["--impure"]